fix circular import by creating a ``format`` package
