+++ bcdi/experiment/setup.py
@@ -27,6 +27,7 @@ from bcdi.experiment.beamline import create_beamline
 from bcdi.experiment.beamline_factory import BeamlineGoniometer, BeamlineSaxs
 from bcdi.experiment.detector import Detector, create_detector
 from bcdi.graph import graph_utils as gu
+import bcdi.utils.format as fmt
 from bcdi.utils import utilities as util
 from bcdi.utils import validation as valid
 from bcdi.utils.io_helper import ContextFile
@@ -697,7 +698,7 @@ class Setup:
 
     def __repr__(self):
         """Representation string of the Setup instance."""
-        return util.create_repr(self, Setup)
+        return fmt.create_repr(self, Setup)
 
     def calc_qvalues_xrutils(self, hxrd, nb_frames, **kwargs):
         """
