+++ bcdi/graph/colormap.py
@@ -16,7 +16,7 @@ import colorcet as cc
 from matplotlib.colors import Colormap, LinearSegmentedColormap, ListedColormap
 
 from bcdi.graph.turbo_colormap import turbo_colormap_data
-from bcdi.utils import utilities as util
+import bcdi.utils.format as fmt
 from bcdi.utils.validation import is_float
 
 data_table = (
@@ -81,4 +81,4 @@ class ColormapFactory:
 
     def __repr__(self):
         """Representation string of the ColormapFactory instance."""
-        return util.create_repr(self, ColormapFactory)
+        return fmt.create_repr(self, ColormapFactory)
