+++ bcdi/utils/parser.py
@@ -18,8 +18,8 @@ from typing import Any, Dict, List, Optional, Type
 
 import yaml
 
+import bcdi.utils.format as fmt
 import bcdi.utils.validation as valid
-from bcdi.utils import utilities as util
 from bcdi.utils.parameters import valid_param
 
 logger = logging.getLogger(__name__)
@@ -280,4 +280,4 @@ class ConfigParser:
 
     def __repr__(self):
         """Representation string of the ConfigParser instance."""
-        return util.create_repr(self, ConfigParser)
+        return fmt.create_repr(self, ConfigParser)
